Fixed several bugs and improved memory usage
Improved memory usage on Aplite Pebbles (original Pebbles). Should not have any crashing now.
Fixed a few other bugs like when it turned on alarms after going into the settings even though they were disabled and not properly resetting the alarm when there was only an alarm for one day of the week.